+++ ks3/key.py
@@ -4,10 +4,12 @@ import hashlib
 import math
 import mimetypes
 import os
-import urllib
 import re
 from hashlib import md5
-
+try:
+    import urllib.parse as parse
+except ImportError:
+    import urllib as parse
 import binascii
 
 import errno
@@ -82,9 +84,9 @@ class Key(object):
 
     def __repr__(self):
         if self.bucket:
-            name = u'<Key: %s,%s>' % (self.bucket.name, self.name)
+            name = '<Key: %s,%s>' % (self.bucket.name, self.name)
         else:
-            name = u'<Key: None,%s>' % self.name
+            name = '<Key: None,%s>' % self.name
 
         # Encode to bytes for Python 2 to prevent display decoding issues
         if not isinstance(name, str):
@@ -349,8 +351,8 @@ class Key(object):
                 fp = EncryptFp(fp, crypt_context, "put")
                 md5_generator = hashlib.md5()
                 md5_generator.update(crypt_context.key)
-                headers["x-kss-meta-key"] = base64.b64encode(md5_generator.hexdigest())
-                headers["x-kss-meta-iv"] = base64.b64encode(crypt_context.first_iv)
+                headers["x-kss-meta-key"] = base64.b64encode(md5_generator.hexdigest().encode()).decode()
+                headers["x-kss-meta-iv"] = base64.b64encode(crypt_context.first_iv).decode()
             if crypt_context.action_info == "upload_part":
                 if crypt_context.part_num == 1:
                     fp = EncryptFp(fp, crypt_context, "upload_part",isUploadFirstPart=True, isUploadLastPart=crypt_context.is_last_part)
@@ -369,7 +371,6 @@ class Key(object):
             else:
                 headers['Content-Length'] = str(len(fp))
             headers["x-kss-meta-unencrypted-content-length"] = str(self.size)
-
             resp = self.bucket.connection.make_request(
                 'PUT',
                 self.bucket.name,
@@ -562,6 +563,7 @@ class Key(object):
             if not replace:
                 if self.bucket.lookup(self.name):
                     return
+            from ks3.encryption import Crypts
             if self.bucket.connection.local_encrypt and self.size:
                 if not crypt_context:
                     crypt_context = Crypts(self.bucket.connection.key)
@@ -604,7 +606,7 @@ class Key(object):
         parameters.
         """
         if not isinstance(string_data, bytes):
-            string_data = string_data.encode("utf-8")
+            string_data = string_data.encode('utf-8')
         fp = BytesIO(string_data)
         r = self.set_contents_from_file(fp, headers, replace, cb, num_cb,
                                         policy, md5, reduced_redundancy,
@@ -652,7 +654,7 @@ class Key(object):
         if response_headers:
             for key in response_headers:
                 query_args.append('%s=%s' % (
-                    key, urllib.quote(response_headers[key])))
+                    key, parse.quote(response_headers[key])))
         query_args = '&'.join(query_args)
         self.open('r', headers, query_args=query_args,
                   override_num_retries=override_num_retries)
@@ -679,15 +681,15 @@ class Key(object):
             counter = 1
             last_iv = ""
             total_part = math.ceil(float(self.size)/self.BufferSize)
-            for bytes in self:
+            for byte in self:
                 if self.bucket.connection.local_encrypt:
                     provider = self.bucket.connection.provider
                     user_key = self.bucket.connection.key
                     crypt_handler = Crypts(user_key)
                     if counter == 1:
                         # For first block, drop first 16 bytes(the subjoin iv).
-                        local_iv = bytes[:crypt_handler.block_size]
-                        bytes = bytes[crypt_handler.block_size:]
+                        local_iv = byte[:crypt_handler.block_size]
+                        byte = byte[crypt_handler.block_size:]
                         server_iv = self.user_meta[provider.metadata_prefix+"iv"]
                         server_iv = base64.b64decode(server_iv)
                         if server_iv and local_iv != server_iv:
@@ -695,23 +697,25 @@ class Key(object):
                         user_iv=local_iv
                     else:
                         user_iv = last_iv
-                    last_iv = bytes[-crypt_handler.block_size:]
+                    last_iv = byte[-crypt_handler.block_size:]
                     if counter == total_part:
                         # Special process of the last part with check code appending to it's end.
-                        full_content = crypt_handler.decrypt(bytes,user_iv)
+                        full_content = crypt_handler.decrypt(byte,user_iv).decode()
                         pad_content_char = full_content[-1]
                         for key in crypt_handler.pad_dict:
                             if crypt_handler.pad_dict[key] == pad_content_char:
                                 pad_content_char = key
                         decrypt = full_content[:-int(pad_content_char)]
                     else:
-                        decrypt = crypt_handler.decrypt(bytes, user_iv)
-                    bytes = decrypt
+                        decrypt = crypt_handler.decrypt(byte, user_iv)
+                    byte = decrypt
                     counter += 1
-                fp.write(bytes)
-                data_len += len(bytes)
+                if not isinstance(byte, bytes):
+                    byte = byte.encode()
+                fp.write(byte)
+                data_len += len(byte)
                 for alg in digesters:
-                    digesters[alg].update(bytes)
+                    digesters[alg].update(byte)
                 if cb:
                     if cb_size > 0 and data_len >= cb_size:
                         break
@@ -752,7 +756,7 @@ class Key(object):
             response_headers = self.resp.msg
             #self.metadata = boto.utils.get_aws_metadata(response_headers,
             #                                            provider)
-            for name, value in response_headers.items():
+            for name, value in list(response_headers.items()):
                 # To get correct size for Range GETs, use Content-Range
                 # header if one was returned. If not, use Content-Length
                 # header.
@@ -947,7 +951,7 @@ class Key(object):
         #print response.msg
         provider = self.bucket.connection.provider
         user_key_reg = provider.metadata_prefix+".*"
-        for header_key, header_value in response.msg.items():
+        for header_key, header_value in list(response.msg.items()):
             reg_match = re.match(user_key_reg, header_key)
             if reg_match:
                 self.user_meta[header_key] = header_value